% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NodeLinkSplitter.r
\name{convertToNodeLink}
\alias{convertToNodeLink}
\title{convertToNodeLink}
\usage{
convertToNodeLink(data, vertices)
}
\arguments{
\item{data}{Edgelist}

\item{Character}{vector specifying the columns used to denote vertices}
}
\value{
A list with the nodes data.frame and links data.frame
}
\description{
Given an edge list split it into nodes and links
}
\details{
Given an edge list split it into nodes and links
}
\examples{
\dontrun{
convertToNodeLink(textExtract, c('One', 'Two'))
}
}
\author{
Jared P. Lander
}

