% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairExtraction.r
\name{dbScroll}
\alias{dbScroll}
\alias{dbScroll.es_conn}
\title{dbScroll}
\usage{
dbScroll(db, nerModel, ...)

\method{dbScroll}{es_conn}(db, nerModel, index = NULL, q = NULL,
  scrollHold = "5m", size = 10, maxNull = 5, ...)
}
\arguments{
\item{db}{database connection}

\item{nerModel}{A ner model supplied by MITIE}

\item{index}{Document index}

\item{q}{Search query}

\item{scrollHold}{Time to hold open scroll state}

\item{size}{Size of entry per shard}

\item{maxNull}{Maximum number of blank returns}

\item{\dots}{Further arguments}
}
\value{
Edgelist of paired entities
}
\description{
Elastic search function
}
\details{
Searches elasticsearch and forms edgelsist of paired entities
}
\author{
Jared P. Lander

Jared P. Lander
}

