% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairExtraction.r
\name{dbSearch}
\alias{dbScroll}
\alias{dbSearch}
\alias{dbSearch.es_conn}
\title{dbSearch}
\usage{
dbSearch(db, nerModel, ...)

\method{dbSearch}{es_conn}(db, nerModel, index = NULL, type = NULL,
  id = NULL, q = NULL, size = 10, ...)

dbScroll(db, nerModel, ...)
}
\arguments{
\item{db}{database connection}

\item{nerModel}{A ner model supplied by MITIE}

\item{index}{Document index}

\item{type}{Type of document}

\item{id}{ID to search}

\item{q}{Search query}

\item{\dots}{Further arguments}

\item{db}{database connection}

\item{nerModel}{A ner model supplied by MITIE}

\item{\dots}{Further arguments}
}
\value{
Edgelist of paired entities

Edgelist of paired entities
}
\description{
Elastic search function

Elastic search function
}
\details{
Searches elasticsearch and forms edgelsist of paired entities

Searches elasticsearch and forms edgelsist of paired entities
}
\author{
Jared P. Lander

Jared P. Lander

Jared P. Lander
}

