% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairExtraction.r
\name{dbScroll.es_conn}
\alias{dbScroll.es_conn}
\title{dbScroll.es_conn}
\usage{
\method{dbScroll}{es_conn}(db, nerModel, index = NULL, scrollHold = "5m",
  size = 10, ...)
}
\arguments{
\item{index}{Document index}

\item{scrollHold}{Time to hold open scroll state}

\item{size}{Size of entry per shard}
}
\author{
Jared P. Lander
}

