% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairExtraction.r
\name{elasticToTbl}
\alias{elasticToTbl}
\title{elasticToTbl}
\usage{
elasticToTbl(data, group = "filename", text = "content")
}
\arguments{
\item{data}{Result from calling an ES search}

\item{group}{Name of grouping variable}

\item{text}{Name of text/content variable}
}
\description{
Converts a list of elastic hits into a tbl
}
\details{
Iterates over a list of results and returns the grouping variable and text (content) variable
}
\author{
Jared P. Lander
}

