% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PairExtraction.r
\name{extractPairs.character}
\alias{extractPairs.character}
\alias{extractPairs.tbl}
\title{extractPairs.character}
\usage{
\method{extractPairs}{character}(x, nerModel, ...)

\method{extractPairs}{tbl}(x, nerModel, ...)
}
\description{
extractPairs.character

extractPairs.tbl
}

