% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EntityExtract.r
\name{extractEntity}
\alias{extractEntity}
\title{extractEntity}
\usage{
extractEntity(entities, tokens, tagNames)
}
\arguments{
\item{tokens}{The tokens from text as found by MITIE}

\item{tagNames}{The possible tags from this MITIE model}
}
\value{
A tbl holding entity info, one row per each
}
\description{
For a collection of entities this grabs the information
}
\details{
For a collection of entities this grabs the information
}
\examples{
\dontrun{
ner_model_path <- "mitie/MITIE-models/english/ner_model.dat"
ner <- NamedEntityExtractor$new(ner_model_path)
tag_names <- ner$get_possible_ner_tags()
theFile <- file.path('data', 'NYTimes', dir(file.path('data', 'NYTimes'))[1])
tokens <- mitie_tokenize(samp[1])
extractEntity(entities, tokens, tag_names)
}
}
\author{
Jared P. Lander
}
\seealso{
extractEntityInfo
}

