% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EntityExtract.r
\name{extractEntityInfo}
\alias{extractEntityInfo}
\title{extractEntityInfo}
\usage{
extractEntityInfo(entity, tokens, tagNames)
}
\arguments{
\item{entity}{A single entity as found by MITIE}

\item{tokens}{The tokens from text as found by MITIE}

\item{tagNames}{The possible tags from this MITIE model}
}
\value{
A one-row tbl holding the Entity, Type, Start and Stop point at the tokens
}
\description{
For a single entity this grabs the information
}
\details{
For a single entity this grabs the information
}
\examples{
\dontrun{
ner_model_path <- "mitie/MITIE-models/english/ner_model.dat"
ner <- NamedEntityExtractor$new(ner_model_path)
tag_names <- ner$get_possible_ner_tags()
theFile <- file.path('data', 'NYTimes', dir(file.path('data', 'NYTimes'))[1])
tokens <- mitie_tokenize(samp[1])
extractEntityInfo(entities[[1]], tokens, tag_names)
}
}
\author{
Jared P. Lander
}
\seealso{
extractEntity
}

