% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentenceExtraction.r
\name{extractSentenceInfo}
\alias{extractSentenceInfo}
\title{extractSentenceInfo}
\usage{
extractSentenceInfo(sentence, nerModel, requiredTerms = NULL,
  ignore.case = TRUE)
}
\arguments{
\item{sentence}{A sentence of text}

\item{nerModel}{A ner model supplied by MITIE}

\item{requiredTerms}{A vector of terms that must be extracted if they exist}

\item{ignore.case}{Logical indicating if requiredTerms is not case sensitive}
}
\value{
A tbl with one row per entity found in the sentence
}
\description{
Given a sentence and a ner model get info on the all the entities found in the sentence
}
\details{
Given a sentence and a ner model get info on the all the entities found in the sentence
}
\examples{
\dontrun{
ner_model_path <- "tests/data/ner_model.dat"
ner <- NamedEntityExtractor$new(ner_model_path)
theText <- "A network of new super PACs said Wednesday that it had raised $31 million to support Senator Ted Cruz's presidential campaign, a sum that could upend expectations in the race for the Republican nomination and rewrite the political rule book for outside spending"
extractSentenceInfo(theText, nerModel=ner)
extractSentenceInfo(theText, nerModel=ner, requiredTerms=c('network'))
}

}
\author{
Jared P. Lander
}
\seealso{
extractEntity
}

