% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TextExtraction.r
\name{extractTextInfo}
\alias{extractTextInfo}
\title{extractTextInfo}
\usage{
extractTextInfo(text, nerModel)
}
\arguments{
\item{text}{A vector of text}

\item{nerModel}{A ner model supplied by MITIE}
}
\value{
A tbl of entity pairs demarcated by the sentence they are paired in.  Could result in multiple combinations.
}
\description{
Given a vector of text, detect sentence-wise entity pairings.
}
\details{
Given a vector of text, detect sentence-wise entity pairings.  Works even if senetences are broken up over multiple elements of the vector of text.
}
\examples{
\dontrun{
ner_model_path <- "mitie/MITIE-models/english/ner_model.dat"
ner <- NamedEntityExtractor$new(ner_model_path)
textFiles <- file.path('data', 'NYTimes', dir('data/NYTimes/'))
thisText <- readLines(textFiles[1])
extractTextInfo(thisText, nerModel=ner)
}
}
\author{
Jared P. Lander
}

