% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteDB.r
\name{filesToDB}
\alias{filesToDB}
\title{filesToDB}
\usage{
filesToDB(files, db, class, type = c("document", "graph"))
}
\arguments{
\item{files}{A vector of file names}

\item{db}{A database connection}

\item{class}{The database class to insert into}

\item{type}{The type of database being written to}
}
\value{
A list of httr response objects
}
\description{
Writes text files to an OrientDB database
}
\details{
Writes text files to an OrientDB database.  Inserts the class, text, file name, time and user into the database object for each file.
}
\examples{
\dontrun{
library(OrientExpress)
theDB <- dbInfo('localhost', database='DocStore', username='TheUser', password='ThePass', port=2480)
filesToDB(c('file1.txt', 'file2.txt'), db=theDB, class='Document')
}

}
\author{
Jared P. Lander
}

