% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TextExtraction.r
\name{getSentencesFromText}
\alias{getSentencesFromText}
\title{getSentencesFromText}
\usage{
getSentencesFromText(text)
}
\arguments{
\item{text}{Text to be split into sentences}
}
\value{
A data.frame with one row per sentence detected
}
\description{
Given text, this extracts all the sentences based on \code{\link[qdaplite]{sent_detect}}.
}
\details{
Given text, this extracts all the sentences based on \code{\link[qdaplite]{sent_detect}}.
}
\examples{
getSentencesFromText(c('This is one sentence in a single element.', 'This element will hold multiple sentences. This is the second. And this is the third and last.'))
}
\author{
Jared P. Lander
}

