% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EntityPairing.r
\name{pairEntities}
\alias{pairEntities}
\title{pairEntities}
\usage{
pairEntities(data, entity = "Entity")
}
\arguments{
\item{data}{A tbl created by extractSentenceInfo}

\item{The}{name of the column holding entity}
}
\value{
A two-column tbl like a pair list between the entities found together
}
\description{
Given a list of entities find all combinations.
}
\details{
Given a list of entities find all combinations.  This expects the data to be of the form |Entity|Type|Start|Stop| as created by extractSenetenceInfo.
}
\examples{
\dontrun{
ner_model_path <- "mitie/MITIE-models/english/ner_model.dat"
ner <- NamedEntityExtractor$new(ner_model_path)
textFiles <- file.path('data', 'NYTimes', dir('data/NYTimes/'))
thisText <- readLines(textFiles[1])
sentences <- sent_detect(thisText)
# run on just one sentence
oneSent <- extractSentenceInfo(sentences[1], nerModel=ner)
pairEntities(oneSent)
}
}
\author{
Jared P. Lander
}
\seealso{
extractSentenceInfo
}

