% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceEdgeList.r
\name{plotForce}
\alias{plotForce}
\title{plotForce}
\usage{
plotForce(data, vertices = c("One", "Two"))
}
\arguments{
\item{data}{An edgelist in data.frame form}

\item{vertices}{Name of edge columns}
}
\value{
A forceNetwork object
}
\description{
Given an edgelist plots a network using forceNetwork from networkD3
}
\details{
Given an edgelist plots a network using forceNetwork from networkD3
}
\examples{
\dontrun{
plotForce(textExtract)
}

}
\author{
Jared P. Lander
}

