% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteDB.r
\name{textFileToDB}
\alias{textFileToDB}
\title{textFileToDB}
\usage{
textFileToDB(file, db, class, type = c("document", "graph"))
}
\arguments{
\item{file}{Path to a text file}

\item{db}{A database connection}

\item{class}{The database class to insert into}

\item{type}{The type of database being written to}
}
\value{
An httr response object
}
\description{
Writes a text file to an OrientDB database
}
\details{
Writes a text file to an OrientDB database.  Inserts the class, text, file name, time and user into the database object.
}
\examples{
\dontrun{
library(OrientExpress)
theDB <- dbInfo('localhost', database='DocStore', username='TheUser', password='ThePass', port=2480)
textFileToDB('A-17.txt', db=theDB, class='Document')
}

}
\author{
Jared P. Lander
}

