% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteDB.r
\name{textToDB}
\alias{textToDB}
\title{textToDB}
\usage{
textToDB(text, db, class, type = c("document", "graph"), filename = NULL)
}
\arguments{
\item{text}{Text to be inserted}

\item{db}{A database connection}

\item{class}{The database class to insert into}

\item{type}{The type of database being written to}

\item{filename}{Name of file that is being inserted}
}
\value{
An httr response object
}
\description{
Writes text to a database
}
\details{
Writes text to an OrientDB database.  Inserts the class, text, file name, time and user into the database object.
}
\examples{
\dontrun{
library(OrientExpress)
theDB <- dbInfo('localhost', database='DocStore', username='TheUser', password='ThePass', port=2480)
theText <- readLines('A-17.txt')
textFileToDB(theText, db=theDB, class='Document')
}
}
\author{
Jared P. Lander
}

