% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EntityExtract.r
\name{extractEntityInfo}
\alias{extractEntityInfo}
\title{extractEntityInfo}
\usage{
extractEntityInfo(entity, tokens, tagNames)
}
\arguments{
\item{entity}{A single entity as found by MITIE}

\item{tokens}{The tokens from text as found by MITIE}

\item{tagNames}{The possible tags from this MITIE model}
}
\value{
A one-row tbl holding the Entity, Type, Start and Stop point at the tokens
}
\description{
For a single entity this grabs the information
}
\details{
For a single entity this grabs the information
}
\examples{
\dontrun{
ner_model_path <- "tests/data/ner_model.dat"
ner <- NamedEntityExtractor$new(ner_model_path)
tag_names <- ner$get_possible_ner_tags()
theText <- "A network of new super PACs said Wednesday that it had raised $31 million to support Senator Ted Cruz's presidential campaign, a sum that could upend expectations in the race for the Republican nomination and rewrite the political rule book for outside spending"
tokens <- mitie_tokenize(theText)
entities <- ner$extract_entities(tokens)
extractEntityInfo(entities[[1]], tokens, tag_names)
extractEntityInfo(entities[[2]], tokens, tag_names)
}
}
\author{
Jared P. Lander
}
\seealso{
extractEntity
}

