% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairExtraction.r
\name{extractPairs}
\alias{extractPairs}
\alias{extractPairs.OrientDB}
\alias{extractPairs.character}
\alias{extractPairs.es_conn}
\alias{extractPairs.tbl}
\title{extractPairs}
\usage{
extractPairs(x, nerModel, ...)

\method{extractPairs}{character}(x, nerModel, ...)

\method{extractPairs}{tbl}(x, nerModel, toSub = NULL, requiredTerms = NULL,
  ignore.case = TRUE, ...)

\method{extractPairs}{OrientDB}(x, nerModel, id = NULL, class, ...)

\method{extractPairs}{es_conn}(x, nerModel, index = NULL, type = NULL,
  id = NULL, q = NULL, search = c("search", "scroll", "page"),
  scrollHold = "5m", size = 10, ...)
}
\arguments{
\item{x}{Either a character vector with names of files or a tbl or database source holding the text.  In the latter cases the text is assumed to be in a column named Text and the grouping variable is assumed to be named File.}

\item{nerModel}{A ner model supplied by MITIE}

\item{toSub}{Named vector where the elements are the pattern and the names are the replacement values}

\item{requiredTerms}{A vector of terms that must be extracted if they exist}

\item{ignore.case}{Logical indicating if requiredTerms is not case sensitive}

\item{id}{List of IDs to read, if \code{NULL} it pulls every record}

\item{class}{The document DB class from which to query}

\item{index}{Document index}

\item{type}{Type of document}

\item{q}{Search query}

\item{search}{Type of search: search, scroll or page}

\item{scrollHold}{Time to hold open scroll state}

\item{size}{Size of entry per shard}

\item{\dots}{Further arguments}
}
\value{
A tbl listing entity cooccurences along with the file name and the sentence number.
}
\description{
Given text, extract entities that cooccur in a sentence
}
\details{
Given text, extract entities that cooccur in a sentence.  Text can be stored in files or in a column in a tbl or database.
}
\examples{
\dontrun{
ner_model_path <- "mitie/MITIE-models/english/ner_model.dat"
ner <- NamedEntityExtractor$new(ner_model_path)
textFiles <- file.path('data', 'NYTimes', dir('data/NYTimes/'))
extractPairs(textFiles, ner)

textDB <- src_sqlite('data/TimesDB/Articles.sqlite3')
textTable <- tbl(textDB, 'Articles')
textTable
extractPairs(textTable, ner)
}

}

