% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SentenceExtraction.r
\name{extractSentenceInfo}
\alias{extractSentenceInfo}
\title{extractSentenceInfo}
\usage{
extractSentenceInfo(sentence, nerModel)
}
\arguments{
\item{sentence}{A sentence of text}

\item{A}{ner model supplied by MITIE}
}
\value{
A tbl with one row per entity found in the sentence
}
\description{
Given a sentence and a ner model get info on the all the entities found in the sentence
}
\details{
Given a sentence and a ner model get info on the all the entities found in the sentence
}
\examples{
\dontrun{
ner_model_path <- "mitie/MITIE-models/english/ner_model.dat"
ner <- NamedEntityExtractor$new(ner_model_path)
textFiles <- file.path('data', 'NYTimes', dir('data/NYTimes/'))
thisText <- readLines(textFiles[1])
sentences <- sent_detect(thisText)
extractSentenceInfo(sentences[1], nerModel=ner)
}
}
\author{
Jared P. Lander
}
\seealso{
extractEntity
}

